ref: Rename ParticipantChannelAllocator.
