+++ src/main/java/org/jitsi/jicofo/conference/colibri/v1/BridgeSession.java
@@ -167,7 +167,7 @@ class BridgeSession
             if (participantInfo != null && participantInfo.getHasColibriSession())
             {
                 terminatedParticipants.add(participant);
-                participant.setChannelAllocator(null);
+                participant.setInviteRunnable(null);
                 participantInfo.setTransport(null);
                 participantInfo.setColibriChannels(null);
                 participantInfo.setHasColibriSession(false);
